Move body content to separate file.
